+++ source/include/Ostap/Bernstein3D.h
@@ -446,6 +446,12 @@ namespace Ostap
       ///  vector  of basic  Bernstein polynomials
       VB m_bz ; //  vector  of basic  Bernstein polynomials
       // ======================================================================
+    private: // some workspace 
+      // ======================================================================
+      mutable std::vector<double> m_fx { 1 , 0.0 } ;
+      mutable std::vector<double> m_fy { 1 , 0.0 } ;      
+      mutable std::vector<double> m_fz { 1 , 0.0 } ;      
+      // ======================================================================      
     } ;
     // ========================================================================
     ///  Bernstein plus      constant
@@ -832,6 +838,12 @@ namespace Ostap
       ///  vector  of basic  Bernstein polynomials
       VB m_b ; //  vector  of basic  Bernstein polynomials
       // ======================================================================
+    private: // some workspace 
+      // ======================================================================
+      mutable std::vector<double> m_fx { 1 , 0.0 } ;
+      mutable std::vector<double> m_fy { 1 , 0.0 } ;      
+      mutable std::vector<double> m_fz { 1 , 0.0 } ;      
+      // ======================================================================      
     } ;
     // ========================================================================
     ///  Bernstein plus      constant
@@ -1228,6 +1240,12 @@ namespace Ostap
       ///  vector  of basic  Bernstein polynomials
       VB m_bz ; //  vector  of basic  Bernstein polynomials
       // ======================================================================
+    private: // some workspace 
+      // ======================================================================
+      mutable std::vector<double> m_fx { 1 , 0.0 } ;
+      mutable std::vector<double> m_fy { 1 , 0.0 } ;      
+      mutable std::vector<double> m_fz { 1 , 0.0 } ;      
+      // ======================================================================      
     } ;
     // ========================================================================
     ///  Bernstein plus      constant